refactor: smart-wallet-utils module
